Add exposure_time_fraction subroutine.
This can be used to obtain the relationship between integration
time and exposure time for each observing mode.  The values
have been taken from the SCUBA-2 Trac, ticket #950.  The same data
are also incorporated in the tA and tB values stored in the module,
but the existing layout of the data has been retained for
consistency with the original ITC.  A test script has been added
which has the same numbers pasted directly from the ticket to
ensure that they have been entered correctly in the code.